+++ app/screens/setting.py
@@ -56,12 +56,12 @@ Builder.load_string('''
                 MDList:
                     OneLineIconListItem:
                         text: "Demo (How to use)"
-                        on_release: app.open_link(self, "https://youtube.com/watch?v=Sbc2sClECdk")
+                        on_release: app.open_link(self, "https://youtube.com/watch?v=UGAxZFuZRtc")
                         IconLeftWidget:
                             icon: "youtube"
                     OneLineIconListItem:
                         text: "Documentation (Blog)"
-                        on_release: app.open_link(self, "https://blog.daslearning.in/llm_ai/ml/ai-cctv.html")
+                        on_release: app.open_link(self, "https://blog.daslearning.in/microcontroller/esp32/automatic-vehicle-indicator.html")
                         IconLeftWidget:
                             icon: "file-document-check"
                     OneLineIconListItem: